Split type predicates into own file; fixed length
- (length) now properly computes the length of a
  proper list, and throws TypeException if the
  data passed in is not a proper list.
- Type predicates besides those for pairs/lists
  now live in lists.hpp.
